Specify default precision and scale for numerics
Two issues here:
1. In the source DB, we often rely on postgres default precision/scale for numerics (https://www.postgresql.org/docs/9.1/datatype-numeric.html)
2. The default precision for numerics is different on redshift (https://www.postgresql.org/docs/9.1/datatype-numeric.html)

Even if the target tables have the right precision/scale for numeric columns, there's still the issue of the temp staging tables needing to have some precision and scale specified in order to prevent rounding to integer during replication. This is a big hammer, but works for our use cases.