Panic rather than returning if there is a messy shutdown
The purpose of this change is to ensure a non-zero exit code in the case of a messy shutdown.

If we were not using Tokio, we would be able to return a Result type from main, but this approach does not work with Tokio.

There are other approaches we could consider here (see: https://docs.rs/quit/0.1.0/quit/index.html and https://www.reddit.com/r/rust/comments/fj7k3b/introducing_quit_exit_cleanly_with_an_exit_code/), but thought I'd start with the simplest.