Add query timeout logic for query that checks if table exists
Uses the tokio timeout functionality to detect when a query has been running longer than expected on the client side. Will attempt to cancel the query if timeout occurs.

This is necessary as, despite the presence of server timeouts, there are cases where connection to the server is lost and the client does not detect it. In those cases, the client will wait indefinitely for the query to complete, which causes a buildup of changes to be processed. This detects such a situation and returns an error result in such a case so that the query can be re-attempted.