Change existence check query
Check for table existence from the source `pg_catalog` tables, rather than via `information_schema` view. The `information_schema.tables` view implementation queries these catalog tables, with the addition that it checks a user's privs against the table in question, which I don't think we need (as the table either exists or it doesn't and re_dms doesn't attempt to tweak privs once created). Our theory by changing this query is that it will potentially be faster and less prone to blocking.