Use `TargetsTablesColumnNames` as a cache for knowing if a table already exists in the target
Previously, on every batch/table, it would run a query to check if the table exists. As re_dms does not drop tables, we can safely assume that if we see a table at startup it will remain in place indefinitely and skip the existence check. This should speed up performance and reduce pressure on the target DB.