[PLT-XXX] Ability to apply database changes in a database transaction
This changes the main "apply" flow to open a database transaction at the outset, and only commit the changes once, at the end of the apply.

The reason we want to do this is to reduce the number of commits performed by re_dms, in order to reduce time spent in the commit queue.

Note that there's a few parameter changes required because the [GenericClient](https://docs.rs/deadpool-postgres/0.14.0/deadpool_postgres/trait.GenericClient.html) trait does not include `cancel_token()` for some odd reason.