[PLT-1337] Re-implement departitioning
This change moves where de-partitioning happens on the `table_name` from `schema_and_table_name` to the actual creation of the `TableName`.

The de-partitioning had to be moved because the old way would cause problems with database serialization (basically, you do not want two transactions concurrently against the same table in Redshift). The issue was that `re_dms` would consider the tables the same physically, in Redshift, but was considering them logically different in re_dms, so they would apply changes from different source (partitioned) tables in parallel, when they should have been applied in a single batch.