[PLT-XXX] Make Docker run on amd64 platform
This tells Docker to run on x86-64 architecture when compiling re_dms for the target system.

Without this, it will target `ARM aarch64` if building on a mac with an arm chip. This won't executable won't be able to run on x86 targets and you get a `Failed to execute command: Exec format error` error when you try.