[PLT-2254] Ability to control whether updates to missing tables are ignored
This provides and ENV, `IGNORE_UPDATES_TO_MISSING_TABLES`, which, when set to `true` will ignore any updates made to missing tables.

We want to be able to configure this behaviour, as in prod environments, this case should never happen, and we want to panic if it does. However, in non-prod envs, we might turn on replication without first doing a backfill, and this case will happen.